修改基础类型

+++ GJLivePush/Video/GJH264Decoder.m
@@ -51,7 +51,7 @@
                                                     (__bridge CFDictionaryRef)(destinationImageBufferAttributes),
                                                     &callBackRecord,
                                                     &_decompressionSession);
-    NSLog(@"Video Decompression Session Create: %@  code:%d  thread:%@", (status == noErr) ? @"successful!" : @"failed...",status,[NSThread currentThread]);
+    NSLog(@"Video Decompression Session Create: %@  code:%d  thread:%@", (status == noErr) ? @"successful!" : @"failed...",(int)status,[NSThread currentThread]);
 }
